Ensure that having no radios does not lead to a crash when we seek to synchronize all radios.
Former-commit-id: 51a13864e0cd2b1a2c47fb395b8da9b1f5b2ff61

+++ CC/Sounder/include/sdr-lib.h
@@ -6,8 +6,8 @@
 #include <cstddef>
 #include <cstdint>
 #include <cstdlib>
-#include <iostream>
 #include <fstream>
+#include <iostream>
 #include <string>
 
 class Radio {
@@ -71,7 +71,7 @@ private:
 
     void radioTrigger(void);
     void sync_delays(int cellIdx);
-    SoapySDR::Device* baseRadio(int cellId);
+    SoapySDR::Device* baseRadio(size_t cellId);
     void collectCSI(bool&);
     static void dciqMinimize(SoapySDR::Device*, SoapySDR::Device*, int, size_t, double, double);
     static void setIQBalance(SoapySDR::Device*, int, size_t, int, int);
